# swagger.yaml
openapi: 3.0.3
info:
  title: BookBazaar
  description: >-
    # BookBazaar

    ## 🏁 Installation

    Follow [these](https://github.com/syprogrammer/bookbazaar) instructions for the installation and project setup.

    ## 🐞 Reporting issues

    If you find API issues or bugs when testing, please create an [issue](https://github.com/syprogrammer/bookbazaar/issues/new) and assign it to a member of the development team.
  version: 1.3.1
servers:
  - url: http://localhost:{port}/api/v1
    description: Development server (local)
    variables:
      port:
        default: 3000
        description: The local server port 
        
  - url: https://bookbazaar.vercel.app/api
    description: Production server
paths:
...
  /books/{bookId}:
    parameters:
      - name: bookId
        in: path
        required: true
        description: The ID of the book
        schema:
          type: string
    patch:
      tags:
        - 📢 Book APIs

      summary: Update book
      description: >-
        The API endpoint allows you to update book in db.

      operationId: updateBook

      responses:
        "200":
          description: Update Book
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: "true"
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: "27925"
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:47:47 GMT
            ETag:
              schema:
                type: string
                example: W/"6d15-0KcGCMfWHfOKgogQFckT/3DtmQw"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
      requestBody:
            required: true
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    title:
                      type: string
                      example: The Hobbit
                    author:
                      type: string
                      example: j.r.r. tolkien
                    genre:
                      type: string
                      example: fantasy
                    bookCoverImage:
                      type: string
                      format: uri
                      example: https://example.com/images/hobbit.jpg
                    description:
                      type: string
                      example: A fantasy adventure that leads to an unexpected journey.
                    price:
                      type: number
                      example: 799
                    quantity:
                      type: integer
                      example: 7
                    publishedYear:
                      type: integer
                      example: 1937
                example:
                  title: The Hobbit
                  author: j.r.r. tolkien
                  genre: fantasy
                  bookCoverImage: https://example.com/images/hobbit.jpg
                  description: A fantasy adventure that leads to an unexpected journey.
                  price: 799
                  quantity: 7
                  publishedYear: 1937
  
    get:
      tags:
        - 📢 Book APIs

      summary: Get book by id
      description: >-
        The API endpoint allows you to get book by id.
      

      operationId: getBookById

      responses:
        "200":
          description: Get Book by id response
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: "true"
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: "27925"
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:47:47 GMT
            ETag:
              schema:
                type: string
                example: W/"6d15-0KcGCMfWHfOKgogQFckT/3DtmQw"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    title:
                      type: string
                      example: The Hobbit
                    author:
                      type: string
                      example: j.r.r. tolkien
                    genre:
                      type: string
                      example: fantasy
                    bookCoverImage:
                      type: string
                      format: uri
                      example: https://example.com/images/hobbit.jpg
                    description:
                      type: string
                      example: A fantasy adventure that leads to an unexpected journey.
                    price:
                      type: number
                      example: 799
                    quantity:
                      type: integer
                      example: 7
                    publishedYear:
                      type: integer
                      example: 1937
                example:
                  title: The Hobbit
                  author: j.r.r. tolkien
                  genre: fantasy
                  bookCoverImage: https://example.com/images/hobbit.jpg
                  description: A fantasy adventure that leads to an unexpected journey.
                  price: 799
                  quantity: 7
                  publishedYear: 1937
    delete:
      tags:
        - 📢 Book APIs

      summary: Delete book
      description: >-
        The API endpoint allows you to Delete book in db.

      operationId: deleteBook

      responses:
        "200":
          description: delete Book
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: "true"
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: "*"
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: "27925"
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:47:47 GMT
            ETag:
              schema:
                type: string
                example: W/"6d15-0KcGCMfWHfOKgogQFckT/3DtmQw"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
...